requirements.txt
Pillow~=9.3.0
StreamDeck~=0.9.3